+++ test/lightning_web/live/workflow_live/edit_test.exs
@@ -1229,7 +1229,9 @@ defmodule LightningWeb.WorkflowLive.EditTest do
         })
         |> render_submit()
 
-      assert html =~ "a workflow with this name already exists in this project."
+      assert html =~
+               "A workflow with this name already exists (possibly pending deletion) in this project."
+
       assert html =~ "Workflow could not be saved"
 
       assert view